[p2p] E2E Connectivity Test + Return Send Success (#52)

+++ cryptography/src/bls12381/benches/dkg_recovery.rs
@@ -12,7 +12,7 @@ fn benchmark_dkg_recovery(c: &mut Criterion) {
                 || {
                     // Create contributors
                     let mut contributors = (0..n)
-                        .map(|i| insecure_signer(i as u16).me())
+                        .map(|i| insecure_signer(i as u64).me())
                         .collect::<Vec<_>>();
                     contributors.sort();
